FIX: did not copy over bins attribute to caller (pvalue plots)
Refactoring in #906 lead to the bins argument being hard-coded at
25 rather than the user-requested value.

Took the advantage of moving the internal _run_pvalue routine
into the only routine that called it. I am also a bit-more explicit
about optional arguments, by naming them, to avoid an issue seen
in at least one case (where a list of optional arguments which were
not named were in a different order to what was expected).