Tests: clean up test_plot_pvalue
Minor clean up of the test code (use default fixture for logging,
avoid double set_source call, note that there's a disconnect in
the test directory and code).

Also extends the test to check that the nums argument is being
recognized. This was added after I realized we are not passing
the bins argument correctly (see forthcoming commit), but it
turns out that we don't use this argument anyway.