Tests: add bins check to plot_pvalue
The bins argument to plot_pvalue is not being sent down to the
right place, so add a check for when it's fixed (so the test
is marked as xfail).

I don't know why the number of bins is one more than requested
(so 21 when you ask for 20 and the current code hardcodes bins=25
so it returns 26 bins).