Tests: add example of #1214
We add a number of related tests of plot_pvalue to identify what
causes issue #1214 and to test out a few things such as using a
model or an image created from that model calculates the same
thing (or, it should).

We need a FITS image for these tests and as it's small we create
it on the fly rather than add it to the test suite (in large part
because we may decide to change the PSF to improve the test and
that's a lot easier to do here than having to update a binary
file).

The tests I added are based on the #1214 report but use different
data (e.g. const + gauss not gauss + gauss) and so it's not obvious
to me if this is a "physically meaningful" test, but it does at
least create repeatable behavior and triggers the bug.

Note that many of the tests fail because we do not handle the
bins argument correctly, which we are about to fix.