Remove support for the empty region
We don't need an "empty" region in Sherpa so it's safer to drop
support for this as it simplifies a number of tests. We can see
that it is only the low-level region tests that are changed;
there's no "actual" sherpa test that uses this feature.

The Region constructor is now more type-safe with the fileflag
argument, casting it to a bool rather than int (the result is
still an int in C/C++ scope but the p flag, added in Python 3.3,
now does the bol conversion to a value of 0 or 1.