FIX: #1214
The problem in #1214 is that the internal state handling for image
objects (used in picking) was relying on undefined (or, perhaps,
undocumented) behavior from the Region class which had been cleaned
up in #968. The old region class allowed you to say Region('None')
and get back an "empty" region filter (in fact any no-valid region
specification would do this) which was triggered by the state
handler converting a None for the region field to 'None', which
would then get passed to Region.

We now are more explicit in the state handler, and check for
the region field being None,  in which case we do nothing.

There is an additional change to support having an "empty"
region string (when region support is turned on) since now
Region('') would fail so we want to support it.