Tests: add tests based on #437
Add two very-simple ssave_all tests for image data: one with no
filter or model and one with a filter and a model.

The first fails because of #437 and so is marked xfail.

The second produces invalid results because of #1245 but is not
marked as a failure (i.e. this tests the current behavior).