FIX: #437
Tweak the serialization code so that for images we do not write
out a filter specification if it is "empty".

We could update notice2d/notice2d_id to treat a filter of ''
as either "do nothing" or "notice everything" which would
allow old save files to be run with Sherpa, but the fix
is easy (remove the problematic line) so I do not do that here.

Note that the filter expression language here is rather opaque
since (e.g. from PHA data) we'd expect an empty filter string
to match no data, but here it doesn't. Perhaps we should change
get_filter to report 'field()' rather than '', but that's a
larger change (and would likely have a number of downstream
consequences).