Switch combine to separate union/subtract methods for Region
The combine method handles both a+b and a-b; let's separate this
into union (a+b) and subtract (a-b). This makes the C++ code a
little clearer in the two cases and makes their use a bit more
explicit in the Python code.

Note that the only changes needed were low-level tests of the
region code.