Correctly union and subtract regions
I had attempted to keep the underlying combination logic when
writing #968 but this appears to have been wrong. This is closer
to how the CIAO region library handles the union and subtract
cases.

You can see that filter expressions which were

    a&b&c

are now

    a|b|c

which makes more sense (& is AND and not OR). It also is clever
enough to fix up some "unnescessary" regions, e.g.

    a-b

becomes

    a

when b does not overlap a, and

    Region(a).union(b).subtract(c)

when c only overlaps a will create

    a&!b|c

One interesting change is that if you start with an "ignore" filter
we used to just to just

    reg = reg.invert()

but we now explicitly add a field definition - that is:

    Region('field()').subtract(reg)

This is because without this the commands

    notice2d()
    ignore2d('circle(100,100,10)')
    ignore2d('circle(140,140,10)')

would create an empty regon filter - this is shown in the test
sherpa/astro/utils/tests/test_astro_data2.py::test_img_get_filter_included_excluded