FIX #1024
The sherpa.astro.plot.ModelHistogram class - used by plot_model but
not plot_fit for PHA data - has, for some reason, the requirement
to plot the model at the "native" resolution of the response, not
the current grouping.

a) we should really make this optional and do it for data and
   models (i.e. move this feature higher up the plot class)
   [but that's not what this commit is about, it's just a pet
    annoyance of mine]

b) To do this, the prepare method temporarily has to ungroup the
   data, but it's not as simple as just calling ungroup, as we
   need to apply the "appropriate" filter on the ungrouped data
   (due to the way we handle the mask at this time; see
   issue #1249). The existing logic for doing this looks sensible
   but it unfortunately broke an invariant we require, but do
   not mandate, that the DataPHA.mask field is used in part to
   identify when we should use the special-case behavior for
   the first call to notice.

The prepare method is now much-more explicit about behavior,
separating out the different classes, rather than having a single
block with multiple if blocks within it, as it makes the intent
clearer (as I know I found this code confusing the first n times
I looked at it, in part because of the complete lack of
documentation).

The fix for #1024 is the fact that we now explicitly handle the
"not filtered" case separately to the "filtered" case, and so
don't end up breaking the mask invariant mentioned above in the
"not filtered" case.

I have also simplified the restoration of the filtered state,
as we don't need to go through the "re-run notice calls that
match the original state" phase when we can just restore the
mask attribute.