Add a module to parse XSPEC model.dat and helper scripts
sherpa.astro.utils.xspec is added to help handle XSPEC model
routines: primarily parsing a model.dat file but also some
helper routines to create the code we need to wrap these models.

This code is pramarily useful for

a) updating the XSPEC module to a new XSPEC release
   - scripts/check_xspec_update.py
     can be used to see what changes are needed
   - scripts/add_xspec_model.py
     will print out the basic code needed to support a new XSPEC
     model
   The idea is to automate the code generation but there are still
   a number of changes that need to be done manually

b) for users who want to use XSPEC user models with Sherpa. This
   functionality is currently provided by the CIAO contributed
   code create_xspec_user_model from which much of this commit
   was taken (i.e. it will simplify that script).