Re-order XSPEC model classes to match model.dat
There is no "correct" way to order the XSPEC model classes, so
this is convention. I have decided to use the order given by the
model.dat file because

- a lot of our models are in that order (since the module was
  originally auto-generated from a model.dat)
- it makes it easier to map between model.dat and this file

I would have also liked to update sherpa/astro/xspec/src/_xspec.cc
to use this ordering but the conditional compilation, and the fact
that the function names are different to the model names, make
this harder to arrange, so I decided it's not worth it.