Update parameter values to match HEASOFT 6.29 (hard limits)
This focuses on the hard limits but a number of switch parameters
have their soft and hard limits removed (to match the model.dat file)
rather than the limits that were given, which were chosen either by
reading the model description (e.g. apec-like switches where it
tells you the values it accepts) or a random range like -100 to 100.

Note that sherpa/astro/tests/test_astro.py::test_proj_bubble is
tweaked here as the covariance results do seem to be very small
compared to the other error analysis routines, so you can see
quite-large changes in the covar results.