Update parameter values to match HEASOFT 6.29 (not hard limits)
This fixes most of the parameter differences between the HEASOFT
6.29 model.dat file and our models, excluding hard limits and
changing the nH units field from '10^22 atoms / cm^-2' to '10^22'
as that changes a number of tests (serialization) and likely will
impact the CIAO regression tests (when a number of other changes
are happening to). The nH units change could be made at a later
time.