Models: add create method for composite models
The idea is that you can create a new version of the composite
model expression. This is currently unused but will be used.

The aim is to make it easy to create a new unary/binary op
instance with a different model (or models). The existing
API doesn't make this cleanly available, so we add the
create methods.

They could have only been added to Unary/BinaryOpModel,
rather than CompositeModel. Let's see if this is a good idea.