Add expand/separate routines
We can now separate out a model expression into it's separate
"additive" components:

  a * (b + c) -> (a * b) + (a * c)

Although we rarely use operators outside of addition or
multiplication this attempts to support them all. To support
subtraction we will "move" things like

  a - b

to

  a + (-b)

This can lead to model expressions like "-(-(b))" which we could
try to clean up, but it is not worth the complication with the
expected use case of these routines.