Initial XSPEC 12.12.0 support
This only supports building against XSPEC 12.12.0, it does not
add any new models from this release, or take advantage of any
changes in it.

Thanks to the header reorganization we now have to import
xsFortran and funcWrappers from a different place, but fortunately
the XSPEC version defines @olaurino added make this easy.

I have noted that we should update from the xsFortran to C++
FunctionUtility interface, as it has more features, but that is
for a different time.

The idea here is to check that these changes do not break old
XSPEC builds before adding new functionality.