XSPEC parameters can now have their hard limits changed
Instead of allowing a parameter value to be set outside the hard
limits, with it's attendant code complications, we now follow
XSPEC and allow the XSParameter class to change the hard limits
of the model.

It is now up to the user to ensure that the parameter is frozen.

Example:

    mdl.fpl.set(hard_min=-1, val=-1, frozen=True)

or

    mdl.fpl.hard_min = -1
    mdl.fpl.val = -1
    mdl.fpl.freeze()