Address @hamogu review comments and improve error testing
The namefunc argument has been changed from None to add_xs_prefix.

A number of tests have been made to check the error handling. Note
that the description of the file in

http://heasarc.gsfc.nasa.gov/xanadu/xspec/manual/XSappendixLocal.html

does not fully describe the behavior of the code (as can be seen by
looking at the XSPEC model.dat file) so some of the parsing has
been written to match reality.