Remove special-case of first filter when outside the data range
The DataPHA class now behaves like the Data1D (and other) classes
do when the first noticed range lies outside the independent axis:
it filters out all data. That is

    pha.units = 'channel'
    pha.notice()
    pha.notice(lo=2000)

when the data covers the channel range of 1 to 1024 will result in
a filter that removes all data - and hence

    pha.get_filter() == ''

Prior to this commit we had added a special case so that

    pha.get_filter() == '1:1024'  # or whatever the full data range is

It did not make sense to have this special case on top of a special
case (the handling of the first notice call).

Note that the only tests that change are those explicitly set up
to check this behavior (i.e. we don't have any "general" tests that
are affected by this change).