Update tests that were changed by #1246
Serialization:

This test was added in #1247 which was merged after I had created
PR #1246, and I never did a rebase during #1246.

Note that the behavior of this test nw makes sense (i.e. you
get consistent behavior).

Data:

This test was added in #1244 which was created after #1246 was
created. This is more a regression-style test, but I argue the output
again makes more sense (but there's not a huge change here).