Update XSkerrconv model for #1275
When I added the XSkerrconv model I had documented the first two
parameters as Index1/Index2 - taken from the model.dat file - but
the actual parameters were labelled Index/Index1. I have renamed
the parameters to match XSPEC. This is a breaking change but I
do not expect many people to be using this model at this time.