Docs: fix get_filter examples
Somehow some of the examples I added were borked, in part due to
all the work done to split up the code into separate PRs and then
the endless rebases.

If we had doctest set up we would have caught this. The trouble
is that setting up doctest is more work than I have time, energy,
or interest in...