GH: clean up scripts
Remove the docs setup from the setup code as we no-longer build
the docs as part of the GH workflows. The docs are built by the
readthedocs system instead. Note that we may decide to enable
doc building with tox, but that would be done differently to the
code we've just removed.

Change the DS9 URL from http to https