GH: minor clean up of workflows
The build instructions are made more atomic, in the sense
that a few steps are now split up so that it's easier to
see what is being run at each stage.

Mark every step that needs it with "shell: bash" (this isn't
technically needed but I think it makes it clearer when we're
running "generic" code via run - and hence we do care about
the shell - versus when we are running some form of an action.