Remove _argument_type_error
The _argument_type_error routine just raises a particular error.
It doesn't add to substantial code savings and is actually makes
the code a bit-harder to read, since it's not obvious (when
skimming the code) that it raises an error. So I've removed it,
using the expicit raise statement in its place, which allows the
code to use 'raise X() from None' statement in several places.