Fix #16
The problem was that the routine was assuming that the dataset
identifier would be defined for both the _models and _sources
dictionaries when it is very unlikely for this to hold. I think
the old code benefited from lazy evaluation of the if statement
that would hide the problem in some cases.

We now are very explicit about the checks - and run the
source case even if the model case passes, as it just makes the
code easier to read and this is not a time-sensitive routine.