Fix #47
The problem that caused #47 is that the generic "save-a-table"
routines could not handle the extra data stored in the DataPHA
class - for instance scalar values and the lists that are
used to store the background and response ids cause slightly
different problems for the crates and pyfits backends.

The surprising solution for #47 is to try to add a meaning to the
_fields attribute of the Data class. Before this commit the
_fields array is used to determine what attributes are displayed
by the __str__ method, along with a reference to an optional
_extra_fields attribute which no class actually defines.

With this commit we now define the _fields attribute as the set
of columns that are written out by the "save to a file" routines,
and the _extra_fields lists the "non-column" data, such as the
shape attribute for image classes and exposure time and
associated background ids for DataPHA objects. This does mean that
the __str__ output is now different - the ordering of the fields
has changed - but I think that is okay (we could add extra fields
to indicate the string output order, but I don't think it's worth
it). Perhaps surprisingly it also changes the IPython display
of these data objects since that ording relies on the _fields
array and so we no-longer include the attributes that are now
in the _extra_fields tuple. As this is informational I do not think
it's a problem, and the one case that shows the difference in
the tests - DataRMF - does not really lose information since we
repeat that information (the DETCHANS setting) in the metadata
listing anyway.