Remove `point` plotting in backend
The definition of the `point` method in the backend was inconsistent
with the other plotting methods in argument (e.g. it lacked
overplot and clearwindow parameters as well as title, xlabel, ...)
and in parameter names (what was called symbol here, is called marker
elsewhere).
This method is just a thin wrapper around `plot` itself - `plot` supports
scatter plots with symbols, so it can be removed.