Tests: improve coverage of error conditions and f-string
Check that a number of corner-case errors are tested. This was
found by checking the test coverage but not with all options
set, so a few of these may have existing checks (e.g. with
X-SPEC installed).

This also addresses some of the recent f-string changes made
in this PR.