data: astro notebook tweaks
For PHA, ARF, and RMF objects we no longer decide on an arbitrary
position in the _fields setting when creating the "data" section
(only used when there's no plot backend). Instead we now just
take all the _fields entries (except for 'name') as the "extra"
information - which shoud be included in the Metadeta section -
is now defined by the _extra_fields attribute.