Fix #143
The fix is easy - change

    if name not in hdu.columns.names:
        return None

    col = hdu.data.field(name)

to

    try:
        col = hdu.data.field(name)
    except KeyError:
        return None

since the field access in the latter is case insensitive. The
test I added now passes.

Note that there are two other routines where I've made the same
change but that we do not have data sets that we can easily use
for testing this. One solutin wuold be to add a grating PI
file (with vector BACKSCAL/SCUP/SCDN or AREASCAL values) where
the column names are lower case, but I didn't want to go to the
effort for this PR (I'd rather wait and review files for when
we add a "proper" astropy backend). At least one of the changes
has been validated with a test.