Testing: add a note about requires_fits decorator
The requires_fits decorator is "interesting", since it checks if
we have any of the supported backend modules available, rather
than checking what the options.io_pkg setting is.

With the recent change in #1185 we can now have a "dummy" backend
selected, and if we do so then any test labelled by
requires_fits is likely to fail. So, long-term we need to decide
what to do in situations like this: examples include

- error out of the tests early, with some sort of "unsupported
  use case" error [not friendly]
- tweak requires_fits so that it checks what the io_pkg setting
  is [not ideal as we may have changed things during testing,
  but maybe okay]
- change the I/O tests to (now we can) switch the I/O backend
  to the supported backends for each test.

However, this is outside the scope of this PR. I just wanted to
add a note about it.

I also made a few style changes to the module

- removed unused imports
- explicit execpt handling
- use of f string