Tests: add some explicit PHA output-format checks
These are low-level tests of the FITS file we create for PHA
objects. This requires direct use of the backend to validate this
information since the Sherpa I/O layer does not provide this
level of detail.

It turns out we can create a PHA file we can't read in with
crates.

The tests are written to include an example of issue #488 -
creating a DataPHA object with "invalid" data types leads to those
types being written out (for, e.g. CHANNEL, QUALITY).