Ensure PHA columns are written out as int/float (fix #488)
I believe that the DataPHA class should enforce the type constraints
on the columns, but there may need to be some work in the I/O layer
anyway so do this first, as it is a lot-less invasive a change.

Unfortunately it's hard to explicitly check the output format when
using crates as it auto-converts int32 to int64.