Take the TELESCOP/INSTRUME/FILTER keywords from response/background fix #1209
XSPEC checks that the TELESCOP/INSTRUME/FILTER keywords match
the PHA file when adding the response or background data. If there
is a difference you see messages like

Warning: RMF TELESCOPE keyword (CHANDRA) is not consistent with spectrum (none)
Warning: RMF INSTRUMENT keyword (ACIS) is not consistent with spectrum (none)
***Warning: Telescope name mismatch between spectrum and  background/correction files
 Source: none Background/Correction ACIS
***Warning: Instrument name mismatch between spectrum and  background/correction files
 Source: none Background/Correction ACIS

We currently do not add this check (we could). Instead, what we
do is that if the DataPHA header has the "default" values for these
header keywords - that is, "none" - then we take these settings
from the first background or response file (ARF or RMF) that is
added to the data. This is to address issue #1209 - using data
created by fake_pha in XSPEC.