data: tweak handling of plot_fac and rate settings
These values are mainly used for plotting, but are set in the
DataPHA class. There are changes to

- slightly-better document them
- add validation of the values

Note that this could change the behavior of old code, for
users setting the type argument of sherpa.astro.ui.set_analysis,
but the old code was inconsistent - e.g. allowing any value that
starts with "rate" but then only checking for "rate" - so this
makes things easier to understand and document. The existing
documentation lists the only valid arguments as "rate" and "counts",
so the code now matches that.