Define API for plotting backend objects
A test is added to check issues found whilst rebasing this work (as part
of #1382).

Part of this removes the ratioline and xaxis options from the plot:

The options have very special use - both together are used only three times in the Sherpa codebase.
They draw a horizontal line. So instead of making them an option to plot
the plot function in the calling plots can just draw a horizontal line
directly. That makes for a cleaner API everywhere.
Users should not see this change, unless they manually changed the preferences for
those lines. In this commit,
I've taken away the possibility to modify a preference to NOT draw those
lines, but I don't see any use case for that.
I think the fact that we have such a prefernece is more an
artifact of how it was implemented (every paraemter automatically
becomes an entry in the XXX_prefs dictionary) then a real decision to have that
as a preferences.

closes #1358

The linecolor option for OrderPlot is now color (the use of
linecolor has been deprecated for a while) and support for
list-of-colors has been added (there are other places this would
be useful but that is being left for later work).

Some of the backend docs are autogenerated using a
new helper function.

Since the plot backend can now be changed during the lifetime of a
program, we need to use a backend-agnostic set of preference settings
for the astro plots. This has a downside, in that a user can no-longer
find out what settings are allowed (that are backend specific), but that
is a "generic" problem, and this change makes the astro code behave
like the other plots.

Co-authored-by: wmclaugh <wmclaugh@cfa.harvard.edu>