Switch to XSPEC typedefs
The changes to _xspec.cc could be replaced by pulling in
functionMap.h from XSPEC - that is, we could just avoid
having to define these symbols - but this needs XSPEC 12.12.0
to insure that the HEADAS/include/ directory is properly
set up (or you reference the BUILD installation rather than
the INSTALL installation of XSPEC).

This changes the FORTRAN routines from 'int *' to 'int&',
requiring a small code change, and labels a few arguments as
const.

Unfortunately the conda XSPEC build we use for testing (CI)
does not include the funcType.h header, and so we can't just
say #include "funcType.h". Instead I have copied over the
values we need.