Better separation of C++ and Python error handling @dtnguyen2
Changes suggested by @dtnguyen2 and then re-worked by @DougBurke
to add the explicit error cases,

The idea is that in most cases the C++ code throws a specialized
case of std::runtime_error so that the handler routine (which is
called by the Python <-> C++ interface) can identify what Python
error to create. An alternative would be for the error type to
be created at the location the error is recognized - that is
there is a call to PyErr_SetString - so that the error handler
can just return NULL.

The repeated logic of "try an action and then apply one of our
specific handlers" could probably be cleaned up.

It is lkely that the code will not work in all cases with an
old XSPEC (pre 12.10.1) but I expect support for these builds
to be removed soon.

There is on estd::runtime_error left, which is left in the
initializer to PyArgTupleBase and is involved with initializing
the model library. With the way the code is currently written we
will have already seen error messages if the library could not
be initialized (see #1388) so I have not tried to change this
part (it should get handled by the generic catch statement).