Tests: add a check for #1397
The problem is that we are not checking for a subclass of the
Parameter class in the error estimators, so the code just creates
two model classes that are the same except for either using
Parameter or a sub-class of Parameter (that adds no extra
functionality). This is to avoid requiring XSPEC to replicate the
original error.