Tests: check more cases of conf/covar/proj
Ensure we test a good range of functionality of the error-estimate
routines. The idea is to test all of them but since they are all
very light-weight wrappers around the _est_errors routine I do
not enforce this for all cases (as it makes coming up with valid
test anwers harder).

There are several cases of checkibg for error conditions, including
one case where the existing docstrings for conf and co include
a call - e.g. conf(1, [2, 3]) - that is actually invalid. We
explicitly check that this case fails but do not update the
documentation, as that will be done in a later commit.