Avoid deprecation warning from Python 3.9
You can get a deprecatio warning from Python 3.9 about using
a hash value for a seed. This is easily "fixed" by casting
the np.int64 value to an int. The assumption is that as this is
the seed then it shouldn't matter whether a hash is used or
an integer (and whether there's any actual difference).