Minor code cleanup for image data
There are a few changes here

- use "modern" Python such as super()
- rework functions so that they return early if not doing very
  much (e.g. if we have a logical system then we can just return
  rather than have to scan through code we don't care about)
- take advantage of the normalization done when setting the
  coordinate value (so that we can remove some redundant checks
  which can never fire)
- some f-string changes

There are also some comment and documentation updates to better
describe the current behavior.

Technically the filter_region change could change the behavior
of the function, but only if the input data has a type that
when multiplied by a SherpaFloat value does not get converted to
a SherpaFloat value, or for which the type conversion to
SherpaFloat is invalid. Any such type is unlikely to be something
we want to be passing around here.