The coord attribute is now readonly for DataIMG
Previously you could change the coord attribute either directly
or with the set_coord method. Unfortunately they are not the same,
and it makes sense to make coord read-only and require users to
use the set_coord method.

The reason for this is because set_coord updates the independent
axis, converting the data to match the selected system. If a
user directly changed the coord attribute then they would find
that the system was no-longer in a valid state, leading to
potential confusion.

Fortunately nothing in Sherpa directly changed the coord attribute,
so it is only a potential breaking change for user code (but in this
case the Sherpa behavior was not guaranteed to be sensible, so
there is an argument for making this change).