fix 1380
The DataIMG class did not keep the original independent axes, but
would create new values based on the current axes when changing
the coordinate setting. This means that going from

  logical to physical to world to logical

would convert the axes each time, which leads to a loss of data
as the world transformation in particular is lossy. The "logical"
values are then ~1e-11 or so different than originally -
depending on the transform - and so the contour command thinks
the bins are not regular.

We could try fixing this with regularization (e.g. forcing logical
values to be integers) but this is not a good solution for the
general case (when the logical coordinates may not be integers).

The solution is to save the original independent axes and coordinate
setting in the _orig_indep_axis attribute, and then use this when
converting between the different coordinate systems. This trades
numerical accuracy against memory, in that this approach requires
more memory. The original code had an explicit comment that it
was saving memory by always re-creating the independent axes
when the coordinate setting was changed rather than (presumably)
storing the values.

There is some attempt to support loading in old "save" files;
that is those created before this change being loaded into a
version with this change. It is not ideal, since we may have
lost the original data so situations like #1380 could still
happen for this data.

One alternative would be to always regenerate the independent
axes when required, which could save some memory but given
how the data access is (e.g. via get_x0 and then get_x1 and
friends) it may not be particularly efficient.