Tests: basic Data2DInt/DataIMGInt tests
The aim was to show some of the problems from issue #1379 but
then I realised it was useful to have a number of checks (some
of these probably already exist but it's hard to check; as these
are quick to run it should not be a problem to have them
repeated).

I assume that the re-work of the data classes in #614 has broken
the Data2DInt and DataIMGInt classes, but as we do not really
exercise these classes it has not been particularly obvious.

It could be that #1359 could have also caused a problem or two,
but this is only likely to be visible after addressing #1379
since the failures added in these tests are present in the 4.13.0
release of Sherpa (which does not contain #1359).

The code that showed #1379 was about evaluating a Delta2D model
on an integrated dataset (Data2DInt, DataIMGInt) rather than an
un-integrated dataset (Data2D, DataIMG) so add a quick check of
the behavior in both cases, as well as adding a "set integrate
flag to False" version.