Fix show error for Data2DInt
This is part of #1379, and happens because the fields attribute
claims the object has x0lo, x0hi, x1lo, and x1hi - needed to get
create the string representatino - but there are no actual
attributes of this name.

We actually do want to be able to access this data, in a similar
matter to Data1DInt/DataPHA, so attributes are added to access
this, followking what is done for Data1DInt. It's not at all clear
what the intended API for accessing this data is (the Data1Dint
xlo/xhi attributes are labelled as "kept for compatibality" so
we do this here too).