ui: fix the documentation for freeze/thaw
The documentation claimed that calling freeze or thaw would
change all models/parameters, but this is not the case. As the
tests that were just added show, calling freeze() or thaw()
does nothing. So the documentation has been adjusted to reflect
the code.

We **could** change the code, but I think that is somewhat
dangerous (e.g. in multi-dataset fits) since we do not have the
concept od an id for these calls, and we do not want to add one
as that would **greatly** complicate the interface.