conf/covar/proj: error out if model has no thawed parameters
If a model is given - e.g. `conf(mdl)` - but all the parameters of
the model are frozen, previously we ignored that model component,
so the result would be to report all thawed parameters for the
source expression.

This is surprising, and it seems better to error out for this case,
which is what the code now does. There is a test of this behavior
(it has changed name for this commit thanks to the change in
behavior).