Tests: explicit ignore_bad tests
The test test_416_b has been reworked to add additional tests
of the quality handling, and renamed to
  test_pha_ignore_bad_group_quality
and
  test_pha_ignore_bad_quality
to test a range of use cases - e.g. what happens when you filter
and then call ignore_bad when you have grouped+quality or
just quality data. The case when there is no quality is already
checked for in test_pha_ignore_bad_no_quality (it errors out).

The reason why we have two tests is partly because the
interaction of grouping and filtering means we want to check
out both cases, but also because currently the ignore_bad
method explicitly behaves differently if the data is grouped,
so we want to check out what happens.