Tests: fix up test problems shown by pytest 7.0.0
Several tests said

    assert pytest.approx(a, b)

which does not actually compare the a and b values. These have
been changed to

    assert b == pytest.approx(a)

Unfortunately one of the tests fails - test_psf2d_kernel_model -
so it has been marked as xfail as it is not obvious what is
going on here. It looks like I added these problem cases as
part of #861 but they were added as "regression-style" tests
so it's not clear what this problem test was actually checking.

Fixes #1428