data: ensure independent axis has the same component sizes
Ensure that the independent axis is formed from components with
the same size. So, a Data1DInt has xlo and xhi values that are
the same size, and Data2DInt has x0lo, x1lo, x0hi, x1hi all the
same size. The idea is to move the validation up the class
hierarchy as far as possible, although as the IntegratedDataSpace1D
case shows i's not always possible.