data: ensure the dependent axis matches the independent axis
This even holds when a data object is created - e.g.

    Data1D('name', [1, 2, 3], [1, 2])

will now raise a DataErr thanks to the mis-match.

The sherpa.astro.data.DataRMF class is an interesting case, as
it is set up as a Data1DInt-derived class, but the
independent axes (the energy grid) and the dependent axis
(matrix) do not fit into the scheme. That is, the sizes do
not match, so the checks on the y attribute are removed for
this class.

Two I/O tests can now be removed since you can't even create the
problematic data (actually, there are ways to create problem
objects but this will be addressed in a later commit).

The code introduces the concept of a "related" component -
that is, a value such as the dependent axis or the staterror
axis - that is related-to the independent axis (has the same
size). This is not really used in this commit, but will be in
later commits.