Tests: what happens if you set the independent axis to None
This is an edge case since we don't expect this to be a common
use case. There are questions about the meaning of this operation,
such as can we just use a single None value, no matter the number
of components, and what should the return value be? So these
tests should be considered as regression tests, to check the
current behavior.