data: make channel/counts aliases for DataPHA
The channel and counts fields are really aliases for the
independent and dependent (also known as y) fields, so make
this explicit.

Of course, the way the independent axis is handled is not-quite
the same as channel (as one is a tuple and one is not) but we can
handle this with the property setter.