Tests: fixes a logical error in the test
The counts field was being set to the channel data, so fix this.

This change doesn't really affect the result, since all we want
to do is create a DataPHA object which causes the call to fail,
but with planned changes the change here is needed, and is
technically the "correct" thing to do. I also updated the code
to be less generic, and only change those fields that we know
need to be changed, which makes the test more readable.